Restructure into more sensible modules and allow pip install -e .
