+++ requirements.txt
@@ -1,8 +1,6 @@
 cairosvg==2.8.2
 chess==1.11.2
-Flask==3.1.2
 matplotlib==3.10.6
-networkx==3.5
 numpy==1.26.4
 Pillow==11.3.0
 tqdm==4.67.1
