+++ src/ttmpRL/__init__.py
@@ -1,0 +1,3 @@
+from .players import Stockfish, Agent, TTAgent
+from .game import Game
+from .scorer import Scorer
