+++ src/ttmpRL/model/__init__.py
@@ -1,0 +1,4 @@
+from .model import ChessModel
+from .encoder import get_game_state
+from .utils import get_model_path
+from .dataloader import GameDataSequence
