+++ src/ttmpRL/model/utils/__init__.py
@@ -1,0 +1,1 @@
+from .pathing import get_model_path
