+++ src/ttmpRL/players/__init__.py
@@ -1,0 +1,3 @@
+from .stockfish import Stockfish
+from .agent import Agent
+from .ttmpt import TTAgent
