+++ src/ttmpRL/players/agent.py
@@ -1,11 +1,8 @@
 import numpy as np
 
-import chessrl.players.netencoder as netencoder
-
-from game import Game
-from chessrl.agents.player import Player
-from chessrl.agents.model import ChessModel
-from chessrl.utils.dataset import DatasetGame
+from .player import Player
+import ttmpRL.model as model
+from ttmpRL.utils import get_uci_labels
 
 
 class Agent(Player):
@@ -20,11 +17,11 @@ class Agent(Player):
     def __init__(self, color, weights_path=None):
         super().__init__(color)
 
-        self.model = ChessModel(compile_model=True, weights=weights_path)
-        self.move_encodings = netencoder.get_uci_labels()
+        self.model = model.ChessModel(compile_model=True, weights=weights_path)
+        self.move_encodings = get_uci_labels()
         self.uci_dict = {u: i for i, u in enumerate(self.move_encodings)}
 
-    def get_move(self, game:Game) -> str:
+    def get_move(self, game:"Game") -> str:
         """ Finds and returns the best possible move (UCI encoded)
 
         Parameters:
@@ -37,19 +34,19 @@ class Agent(Player):
         Returns:
             str. UCI encoded movement.
         """
-        move = Game.NULL_MOVE
+        move = game.NULL_MOVE
         policy = self.predict_policy(game)
         move = game.get_legal_moves()[np.argmax(policy)]
         return move
 
-    def predict_outcome(self, game:Game) -> float:
+    def predict_outcome(self, game:"Game") -> float:
         """ Predicts the outcome of a game from the current position """
-        game_matr = netencoder.get_game_state(game)
+        game_matr = model.get_game_state(game)
         return self.model.predict(np.expand_dims(game_matr, axis=0))[1][0][0]
 
-    def predict_policy(self, game:Game, mask_legal_moves=True) -> float:
+    def predict_policy(self, game:"Game", mask_legal_moves=True) -> float:
         """ Predict the policy distribution over all possible moves. """
-        game_matr = netencoder.get_game_state(game)
+        game_matr = model.get_game_state(game)
         policy = self.model.predict(np.expand_dims(game_matr, axis=0))[0][0]
         if mask_legal_moves:
             legal_moves = game.get_legal_moves()
@@ -60,32 +57,32 @@ class Agent(Player):
     # Separate the NN model from the Agent class. Rewrite it so that
     # the Agent can also validly take position score predictions from   
     # Stockfish or other engines.
-    def train(self, dataset: DatasetGame,
-              epochs=1, logdir=None, batch_size=1,
-              validation_split=0):
-        """ Trains the model using previous recorded games."""
-        if len(dataset) <= 0:
-            return
+    # def train(self, dataset: DatasetGame,
+    #           epochs=1, logdir=None, batch_size=1,
+    #           validation_split=0):
+    #     """ Trains the model using previous recorded games."""
+    #     if len(dataset) <= 0:
+    #         return
 
-        if validation_split > 0:
-            split_point = len(dataset) - int(validation_split * len(dataset))
+    #     if validation_split > 0:
+    #         split_point = len(dataset) - int(validation_split * len(dataset))
 
-            games_train = DatasetGame(dataset[:split_point])
-            games_val = DatasetGame(dataset[split_point:])
-            val_gen = netencoder.DataGameSequence(games_val,
-                                                  batch_size=batch_size)
-        else:
-            games_train = dataset
-            val_gen = None
+    #         games_train = DatasetGame(dataset[:split_point])
+    #         games_val = DatasetGame(dataset[split_point:])
+    #         val_gen = netencoder.DataGameSequence(games_val,
+    #                                               batch_size=batch_size)
+    #     else:
+    #         games_train = dataset
+    #         val_gen = None
 
-        train_gen = netencoder.DataGameSequence(games_train,
-                                                batch_size=batch_size,
-                                                random_flips=.1)
+    #     train_gen = netencoder.DataGameSequence(games_train,
+    #                                             batch_size=batch_size,
+    #                                             random_flips=.1)
 
-        self.model.train_generator(train_gen,
-                                   epochs=epochs,
-                                   logdir=logdir,
-                                   val_gen=val_gen)
+    #     self.model.train_generator(train_gen,
+    #                                epochs=epochs,
+    #                                logdir=logdir,
+    #                                val_gen=val_gen)
 
     def save(self, path):
         self.model.save_weights(path)
