+++ src/ttmpRL/players/ttmpt.py
@@ -5,20 +5,19 @@ import numpy as np
 from numpy.random import randint
 from tqdm import tqdm
 from timeit import default_timer as timer
-from utils.logger import Logger
 import traceback
 
-from chessrl.utils.dataset import DatasetGame
-from chessrl.agents.agent import Agent
-from game import Game
+from ttmpRL.utils import Logger
+from ttmpRL.dataset import GameDataset
+from .agent import Agent
 
 
 def playout_and_save_game(
-        dataset: DatasetGame,
-        starting_game: Game,
+        dataset: "GameDataset",
+        starting_game: "Game",
         agent: Agent,
         stockfish_bin
-    ) -> DatasetGame:
+    ) -> GameDataset:
     """Get agent to play out game against stockfish, starting from particular
     position, and add to dataset."""
 
@@ -65,14 +64,14 @@ class TTAgent(Agent):
         self.tt_games = tt_games
         self.ttt_iters = ttt_iters
 
-    def get_move(self, game:Game, real_game=True, verbose=False) -> str:
+    def get_move(self, game:"Game", real_game=True, verbose=False) -> str:
         """Perform iterations of predicting games and fine-tuning
         the agent before selecting a move."""
 
         logger = Logger.get_instance()
         timer_start = timer()
 
-        move = Game.NULL_MOVE
+        move = game.NULL_MOVE
 
         tmp_agent = self.predict_and_tune(game)
         policy = tmp_agent.predict_policy(game)
@@ -84,7 +83,7 @@ class TTAgent(Agent):
 
         return move
 
-    def predict_and_tune(self, game:Game) -> Agent:
+    def predict_and_tune(self, game: "Game") -> Agent:
 
         logger = Logger.get_instance()
 
@@ -93,7 +92,7 @@ class TTAgent(Agent):
         logger.info("Making predictions and tuning agent...")
         for r in tqdm(range(self.ttt_iters), desc="Tuning iterations"):
 
-            dataset = DatasetGame()
+            dataset = GameDataset()
 
             # play lots of games from the curent board state onwards
             for _ in tqdm(range(self.tt_games), desc="Predicting games"):
