+++ src/ttmpRL/utils/__init__.py
@@ -1,0 +1,2 @@
+from .logger import Logger
+from .chess import get_uci_labels
